Use sqlite-jdbc by default (but still support H2)
